docs : styled 분리
