+++ src/components/Navigation/NavHeader.tsx
@@ -1,51 +1,14 @@
 import { Link } from "react-router-dom";
-import styled from "styled-components";
 import NavBar from "./NavBar";
 import { useState } from "react";
+import { auth } from "../../firebase";
+import { LogoImg, NavHeaderButton, NavHeaderUl } from "./style";
 
-const NavHeaderUl = styled.ul`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  background-color: var(--poke-dex-red);
-  padding: 0.5rem 1rem;
-  position: relative;
+const currentUser = auth.currentUser;
 
-  li a {
-    text-shadow: rgb(233, 233, 233) 3px 0px 0px, rgb(233, 233, 233) 2.83487px 0.981584px 0px,
-      rgb(233, 233, 233) 2.35766px 1.85511px 0px, rgb(233, 233, 233) 1.62091px 2.52441px 0px,
-      rgb(233, 233, 233) 0.705713px 2.91581px 0px, rgb(233, 233, 233) -0.287171px 2.98622px 0px,
-      rgb(233, 233, 233) -1.24844px 2.72789px 0px, rgb(233, 233, 233) -2.07227px 2.16926px 0px,
-      rgb(233, 233, 233) -2.66798px 1.37182px 0px, rgb(233, 233, 233) -2.96998px 0.42336px 0px,
-      rgb(233, 233, 233) -2.94502px -0.571704px 0px, rgb(233, 233, 233) -2.59586px -1.50383px 0px,
-      rgb(233, 233, 233) -1.96093px -2.27041px 0px, rgb(233, 233, 233) -1.11013px -2.78704px 0px,
-      rgb(233, 233, 233) -0.137119px -2.99686px 0px, rgb(233, 233, 233) 0.850987px -2.87677px 0px,
-      rgb(233, 233, 233) 1.74541px -2.43999px 0px, rgb(233, 233, 233) 2.44769px -1.73459px 0px,
-      rgb(233, 233, 233) 2.88051px -0.838247px 0px;
-  }
-`;
-
-const LogoImg = styled.img`
-  background-color: white;
-  padding: 0.2rem;
-  width: 30px;
-  height: 30px;
-  margin-right: 1rem;
-  border-radius: 10px;
-  border: 2px solid var(--water);
-`;
-
-const NavHeaderButton = styled.button`
-  background-color: rgba(0, 0, 0, 0);
-  border: none;
-  &:hover {
-    cursor: pointer;
-  }
-  img {
-    width: 30px;
-    height: 30px;
-  }
-`;
+const handleLogout = () => {
+  auth.signOut();
+};
 
 function NavHeader() {
   const [showNav, setShowNav] = useState(false);
@@ -62,6 +25,10 @@ function NavHeader() {
           </Link>
         </li>
         <li>
+          {
+            currentUser === null ? <NavHeaderButton><Link to="/login">로그인</Link></NavHeaderButton> :
+              <NavHeaderButton onClick={handleLogout}>로그아웃</NavHeaderButton>
+          }
           <NavHeaderButton onClick={toggleNav}>
             <img src="/image/icon/ball.png" alt="네비 버튼 이미지" />
           </NavHeaderButton>
